+++ plugins/postcss/src/index.ts
@@ -1,4 +1,4 @@
-import type { InputOptions, PluginOptions } from "#/@types/options";
+import type { PluginOptions } from "#/@types/options";
 import type { Plugin } from "#/create";
 
 import { createPlugin } from "#/create";
@@ -6,4 +6,4 @@ import { createPlugin } from "#/create";
 const plugin: Plugin = createPlugin();
 
 export default plugin;
-export type { InputOptions, PluginOptions };
+export type { PluginOptions };